refactor: split app feedback and diagnostics
